Remove OpenOCD warnings
Remove OpenOCD warnings:

DEPRECATED! use 'adapter driver' not 'interface'
Warn : An adapter speed is not selected in the init scripts. OpenOCD will try to run the adapter at the low speed (100 kHz)
Warn : To remove this warnings and achieve reasonable communication speed with the target, set "adapter speed" or "jtag_rclk" in the init scripts.